requirements.txt
numpy>=1.21.0,<1.24.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
salesforce-merlion>=1.1.0